cleanup before chat context
