+++ electron/preload/index.ts
@@ -77,8 +77,6 @@ const electronStore = {
   getChatHistory: createIPCHandler<(chatID: string) => Promise<Chat>>('get-chat-history'),
   getSBCompact: createIPCHandler<() => Promise<boolean>>('get-sb-compact'),
   setSBCompact: createIPCHandler<(isSBCompact: boolean) => Promise<void>>('set-sb-compact'),
-  getDisplayMarkdown: createIPCHandler<() => Promise<boolean>>('get-display-markdown'),
-  setDisplayMarkdown: createIPCHandler<(displayMarkdown: boolean) => Promise<void>>('set-display-markdown'),
   getEditorFlexCenter: createIPCHandler<() => Promise<boolean>>('get-editor-flex-center'),
   setEditorFlexCenter: createIPCHandler<(editorFlexCenter: boolean) => Promise<void>>('set-editor-flex-center'),
   getCurrentOpenTabs: createIPCHandler<() => Promise<Tab[]>>('get-current-open-files'),
